src/pyproject.toml
[tool.poetry]
name = "lab05"
version = "0.0.0"
description = "App for CSCI338"
authors = ["Sarah Van Wart <vanwars@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
aiohttp = "3.8.2"
aiofiles = "^23.2.1"
httpx = "^0.25.0"
pytest = "^7.4.2"
asyncpg = "^0.28.0"
SQLAlchemy = "^2.0.0"
psycopg2 = "^2.9.9"
Faker = "^19.12.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
